structural update
-new debugger class for giving advanced debug information
-new modulemerger class for module handling
-welder is a parent of the modulemerger class
-moved regularexpression methods into welding helper
-added Cmodules for struts to be ignored
-added prewelding check for strut to remove them in the weldment
